refactor(structure): merge extract_f0_rmvpe_dml into extract_f0_print
now all of the f0 extractions are united in the extract_f0_print.py file
